+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes for Elasticsearch Plugin
 
+## Unreleased
+
+### Added
+
+- Added console command to list all aliases and indexes of the configured Elasticsearch cluster. 
+
 ## 1.0.0 - 2022-4-08
 
 ### Added